backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
google-generativeai==0.8.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
Pillow==10.4.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
